feat: add script to evaluate scale dependency

+++ dynamic_fusion/utils/dataset.py
@@ -60,7 +60,7 @@ class CocoTestDataset(Dataset):  # type: ignore
     logger: logging.Logger
     scales: Float[np.ndarray, " N"]
 
-    def __init__(self, dataset_directory: Path, scales_range: Tuple[int, int], threshold: float = 1.4) -> None:
+    def __init__(self, dataset_directory: Path, scales_range: Tuple[float, float], threshold: float = 1.4) -> None:
         self.directory_list = sorted([path for path in dataset_directory.glob("**/*") if path.is_dir()])
         self.logger = logging.getLogger("CocoDataset")
         self.threshold = threshold
